--- src/contexts/products.context.jsx
@@ -1,27 +1,0 @@
-import { createContext,useState,useEffect } from "react";
-import {getCategoriesAndDocuments} from '../utils/firebase/firebase.utils';
-//import { addCollectionAndDocuments } from "../utils/firebase/firebase.utils.js";
-//import PRODUCTS from '../shop-data.json';
-//import SHOP_DATA from '../shop-data.js';
-
-
-export const ProductsContext= createContext({
-    products:[]
-});
-
-export const ProductsProvider =({children})=>{
-   // const [products,setProducts] = useState(PRODUCTS);
-    const [products,setProducts] = useState([]);
-    useEffect(()=>{
-        const getCategoriesMap = async() => {
-            const categoryMap = await getCategoriesAndDocuments();
-            console.log(categoryMap);            
-        }
-        getCategoriesMap();
-    },[]);
-   /*  useEffect(()=>{
-        addCollectionAndDocuments('categories',SHOP_DATA);
-    },[]); */   
-    const value = {products};
-    return (<ProductsContext.Provider value={value}>{children}</ProductsContext.Provider>);
-}
