+++ src/index.js
@@ -4,7 +4,8 @@ import { BrowserRouter } from 'react-router-dom';
 
 import App from './App';
 import {UserProvider} from './contexts/user.context';
-import { ProductsProvider } from './contexts/products.context';
+//import { ProductsProvider } from './contexts/products.context';
+import {CategoriesProvider } from './contexts/categories.context';
 import { CartProvider } from './contexts/cart.context';
 import './index.scss';
 import reportWebVitals from './reportWebVitals';
@@ -14,11 +15,11 @@ root.render(
   <React.StrictMode>
     <BrowserRouter>
       <UserProvider>
-        <ProductsProvider>
+        <CategoriesProvider>
           <CartProvider>
             <App/>
           </CartProvider>          
-        </ProductsProvider>        
+        </CategoriesProvider>        
       </UserProvider>     
     </BrowserRouter>    
   </React.StrictMode>
